py-django/requirements.txt
Django==3.2.19
djaodjin-deployutils==0.10.6
